Break boxFunctions into separate files
